+++ MealPlanner/MealPlanner.csproj
@@ -9,4 +9,9 @@
     <PackageReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Design" Version="2.1.10" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="Pages\Meals\" />
+    <Folder Include="Pages\Recipes\" />
+  </ItemGroup>
+
 </Project>
