+++ MealPlanner/Pages/Entries/Create.cshtml.cs
@@ -7,31 +7,38 @@ using Microsoft.AspNetCore.Mvc.RazorPages;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using MealPlanner.Models;
 
-namespace MealPlanner.Pages.Entries {
-  public class CreateModel : PageModel {
-    private readonly MealPlanner.Models.MealPlannerContext _context;
+namespace MealPlanner.Pages.Entries
+{
+    public class CreateModel : PageModel
+    {
+        private readonly MealPlanner.Models.MealPlannerContext _context;
 
-    public CreateModel( MealPlanner.Models.MealPlannerContext context ) {
-      _context = context;
+        public CreateModel(MealPlanner.Models.MealPlannerContext context)
+        {
+            _context = context;
+        }
+
+        public IActionResult OnGet()
+        {
+        ViewData["MealId"] = new SelectList(_context.Meal, "ID", "ID");
+        ViewData["PlanId"] = new SelectList(_context.Plan, "ID", "ID");
+            return Page();
+        }
+
+        [BindProperty]
+        public PlanEntry PlanEntry { get; set; }
+
+        public async Task<IActionResult> OnPostAsync()
+        {
+            if (!ModelState.IsValid)
+            {
+                return Page();
+            }
+
+            _context.PlanEntry.Add(PlanEntry);
+            await _context.SaveChangesAsync();
+
+            return RedirectToPage("./Index");
+        }
     }
-
-    public IActionResult OnGet( int? planId ) {
-      var id = planId;
-      return Page();
-    }
-
-    [BindProperty] public PlanEntry PlanEntry { get; set; }
-
-    public async Task<IActionResult> OnPostAsync(int planId) {
-      if ( !ModelState.IsValid ) {
-        return Page();
-      }
-
-      PlanEntry.PlanId = planId;
-      _context.PlanEntry.Add( PlanEntry );
-      await _context.SaveChangesAsync();
-
-      return RedirectToPage( "./Index" );
-    }
-  }
 }
