+++ MealPlanner/Pages/Entries/Delete.cshtml.cs
@@ -28,7 +28,9 @@ namespace MealPlanner.Pages.Entries
                 return NotFound();
             }
 
-            PlanEntry = await _context.PlanEntry.FirstOrDefaultAsync(m => m.ID == id);
+            PlanEntry = await _context.PlanEntry
+                .Include(p => p.Meal)
+                .Include(p => p.Plan).FirstOrDefaultAsync(m => m.ID == id);
 
             if (PlanEntry == null)
             {
