+++ MealPlanner/Pages/Entries/Details.cshtml.cs
@@ -27,7 +27,9 @@ namespace MealPlanner.Pages.Entries
                 return NotFound();
             }
 
-            PlanEntry = await _context.PlanEntry.FirstOrDefaultAsync(m => m.ID == id);
+            PlanEntry = await _context.PlanEntry
+                .Include(p => p.Meal)
+                .Include(p => p.Plan).FirstOrDefaultAsync(m => m.ID == id);
 
             if (PlanEntry == null)
             {
