+++ MealPlanner/Pages/Entries/Edit.cshtml
@@ -15,14 +15,19 @@
             <div asp-validation-summary="ModelOnly" class="text-danger"></div>
             <input type="hidden" asp-for="PlanEntry.ID" />
             <div class="form-group">
+                <label asp-for="PlanEntry.PlanId" class="control-label"></label>
+                <select asp-for="PlanEntry.PlanId" class="form-control" asp-items="ViewBag.PlanId"></select>
+                <span asp-validation-for="PlanEntry.PlanId" class="text-danger"></span>
+            </div>
+            <div class="form-group">
+                <label asp-for="PlanEntry.MealId" class="control-label"></label>
+                <select asp-for="PlanEntry.MealId" class="form-control" asp-items="ViewBag.MealId"></select>
+                <span asp-validation-for="PlanEntry.MealId" class="text-danger"></span>
+            </div>
+            <div class="form-group">
                 <label asp-for="PlanEntry.Date" class="control-label"></label>
                 <input asp-for="PlanEntry.Date" class="form-control" />
                 <span asp-validation-for="PlanEntry.Date" class="text-danger"></span>
-            </div>
-            <div class="form-group">
-                <label asp-for="PlanEntry.Notes" class="control-label"></label>
-                <input asp-for="PlanEntry.Notes" class="form-control" />
-                <span asp-validation-for="PlanEntry.Notes" class="text-danger"></span>
             </div>
             <div class="form-group">
                 <input type="submit" value="Save" class="btn btn-default" />
