--- MealPlanner/Pages/Recipes/Create.cshtml.cs
@@ -1,42 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.RazorPages;
-using Microsoft.AspNetCore.Mvc.Rendering;
-using MealPlanner.Models;
-
-namespace MealPlanner.Pages.Recipes
-{
-    public class CreateModel : PageModel
-    {
-        private readonly MealPlanner.Models.MealPlannerContext _context;
-
-        public CreateModel(MealPlanner.Models.MealPlannerContext context)
-        {
-            _context = context;
-        }
-
-        public IActionResult OnGet()
-        {
-            return Page();
-        }
-
-        [BindProperty]
-        public Recipe Recipe { get; set; }
-
-        public async Task<IActionResult> OnPostAsync()
-        {
-            if (!ModelState.IsValid)
-            {
-                return Page();
-            }
-
-            _context.Recipe.Add(Recipe);
-            await _context.SaveChangesAsync();
-
-            return RedirectToPage("./Index");
-        }
-    }
-}
