--- MealPlanner/Pages/Recipes/Delete.cshtml
@@ -1,46 +1,0 @@
-﻿@page
-@model MealPlanner.Pages.Recipes.DeleteModel
-
-@{
-    ViewData["Title"] = "Delete";
-}
-
-<h2>Delete</h2>
-
-<h3>Are you sure you want to delete this?</h3>
-<div>
-    <h4>Recipe</h4>
-    <hr />
-    <dl class="dl-horizontal">
-        <dt>
-            @Html.DisplayNameFor(model => model.Recipe.Reference)
-        </dt>
-        <dd>
-            @Html.DisplayFor(model => model.Recipe.Reference)
-        </dd>
-        <dt>
-            @Html.DisplayNameFor(model => model.Recipe.Title)
-        </dt>
-        <dd>
-            @Html.DisplayFor(model => model.Recipe.Title)
-        </dd>
-        <dt>
-            @Html.DisplayNameFor(model => model.Recipe.Description)
-        </dt>
-        <dd>
-            @Html.DisplayFor(model => model.Recipe.Description)
-        </dd>
-        <dt>
-            @Html.DisplayNameFor(model => model.Recipe.Notes)
-        </dt>
-        <dd>
-            @Html.DisplayFor(model => model.Recipe.Notes)
-        </dd>
-    </dl>
-    
-    <form method="post">
-        <input type="hidden" asp-for="Recipe.ID" />
-        <input type="submit" value="Delete" class="btn btn-default" /> |
-        <a asp-page="./Index">Back to List</a>
-    </form>
-</div>
