--- MealPlanner/Pages/Recipes/Delete.cshtml.cs
@@ -1,58 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.RazorPages;
-using Microsoft.EntityFrameworkCore;
-using MealPlanner.Models;
-
-namespace MealPlanner.Pages.Recipes
-{
-    public class DeleteModel : PageModel
-    {
-        private readonly MealPlanner.Models.MealPlannerContext _context;
-
-        public DeleteModel(MealPlanner.Models.MealPlannerContext context)
-        {
-            _context = context;
-        }
-
-        [BindProperty]
-        public Recipe Recipe { get; set; }
-
-        public async Task<IActionResult> OnGetAsync(int? id)
-        {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
-            Recipe = await _context.Recipe.FirstOrDefaultAsync(m => m.ID == id);
-
-            if (Recipe == null)
-            {
-                return NotFound();
-            }
-            return Page();
-        }
-
-        public async Task<IActionResult> OnPostAsync(int? id)
-        {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
-            Recipe = await _context.Recipe.FindAsync(id);
-
-            if (Recipe != null)
-            {
-                _context.Recipe.Remove(Recipe);
-                await _context.SaveChangesAsync();
-            }
-
-            return RedirectToPage("./Index");
-        }
-    }
-}
