--- MealPlanner/Pages/Recipes/Details.cshtml
@@ -1,43 +1,0 @@
-﻿@page
-@model MealPlanner.Pages.Recipes.DetailsModel
-
-@{
-    ViewData["Title"] = "Details";
-}
-
-<h2>Details</h2>
-
-<div>
-    <h4>Recipe</h4>
-    <hr />
-    <dl class="dl-horizontal">
-        <dt>
-            @Html.DisplayNameFor(model => model.Recipe.Reference)
-        </dt>
-        <dd>
-            @Html.DisplayFor(model => model.Recipe.Reference)
-        </dd>
-        <dt>
-            @Html.DisplayNameFor(model => model.Recipe.Title)
-        </dt>
-        <dd>
-            @Html.DisplayFor(model => model.Recipe.Title)
-        </dd>
-        <dt>
-            @Html.DisplayNameFor(model => model.Recipe.Description)
-        </dt>
-        <dd>
-            @Html.DisplayFor(model => model.Recipe.Description)
-        </dd>
-        <dt>
-            @Html.DisplayNameFor(model => model.Recipe.Notes)
-        </dt>
-        <dd>
-            @Html.DisplayFor(model => model.Recipe.Notes)
-        </dd>
-    </dl>
-</div>
-<div>
-    <a asp-page="./Edit" asp-route-id="@Model.Recipe.ID">Edit</a> |
-    <a asp-page="./Index">Back to List</a>
-</div>
