--- MealPlanner/Pages/Recipes/Details.cshtml.cs
@@ -1,39 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.RazorPages;
-using Microsoft.EntityFrameworkCore;
-using MealPlanner.Models;
-
-namespace MealPlanner.Pages.Recipes
-{
-    public class DetailsModel : PageModel
-    {
-        private readonly MealPlanner.Models.MealPlannerContext _context;
-
-        public DetailsModel(MealPlanner.Models.MealPlannerContext context)
-        {
-            _context = context;
-        }
-
-        public Recipe Recipe { get; set; }
-
-        public async Task<IActionResult> OnGetAsync(int? id)
-        {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
-            Recipe = await _context.Recipe.FirstOrDefaultAsync(m => m.ID == id);
-
-            if (Recipe == null)
-            {
-                return NotFound();
-            }
-            return Page();
-        }
-    }
-}
