--- MealPlanner/Pages/Recipes/Edit.cshtml.cs
@@ -1,74 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.RazorPages;
-using Microsoft.AspNetCore.Mvc.Rendering;
-using Microsoft.EntityFrameworkCore;
-using MealPlanner.Models;
-
-namespace MealPlanner.Pages.Recipes
-{
-    public class EditModel : PageModel
-    {
-        private readonly MealPlanner.Models.MealPlannerContext _context;
-
-        public EditModel(MealPlanner.Models.MealPlannerContext context)
-        {
-            _context = context;
-        }
-
-        [BindProperty]
-        public Recipe Recipe { get; set; }
-
-        public async Task<IActionResult> OnGetAsync(int? id)
-        {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
-            Recipe = await _context.Recipe.FirstOrDefaultAsync(m => m.ID == id);
-
-            if (Recipe == null)
-            {
-                return NotFound();
-            }
-            return Page();
-        }
-
-        public async Task<IActionResult> OnPostAsync()
-        {
-            if (!ModelState.IsValid)
-            {
-                return Page();
-            }
-
-            _context.Attach(Recipe).State = EntityState.Modified;
-
-            try
-            {
-                await _context.SaveChangesAsync();
-            }
-            catch (DbUpdateConcurrencyException)
-            {
-                if (!RecipeExists(Recipe.ID))
-                {
-                    return NotFound();
-                }
-                else
-                {
-                    throw;
-                }
-            }
-
-            return RedirectToPage("./Index");
-        }
-
-        private bool RecipeExists(int id)
-        {
-            return _context.Recipe.Any(e => e.ID == id);
-        }
-    }
-}
