--- MealPlanner/Pages/Recipes/Index.cshtml.cs
@@ -1,28 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.RazorPages;
-using Microsoft.EntityFrameworkCore;
-using MealPlanner.Models;
-
-namespace MealPlanner.Pages.Recipes
-{
-    public class IndexModel : PageModel
-    {
-        private readonly MealPlanner.Models.MealPlannerContext _context;
-
-        public IndexModel(MealPlanner.Models.MealPlannerContext context)
-        {
-            _context = context;
-        }
-
-        public IList<Recipe> Recipe { get;set; }
-
-        public async Task OnGetAsync()
-        {
-            Recipe = await _context.Recipe.ToListAsync();
-        }
-    }
-}
